app/backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1  # Pin to 4.0.1 for passlib compatibility

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0